PIM-6353: Indexing proposition #1
In order to index:
- ProductInterface in both P & PM indexes (in the right index format)
- ProductModelInterface only in PM

This PR suggest to define:

- Two services for the indexing (pointing to the same class ObjectIndexer) capable of
normalizing and indexing the objects in their dedidcated ES index.

- Two separate POST_SAVE subscribers:
  * One service which calls both of the above indexers to index the ProductInterface
  objects in both indexes

  * One service which calls the indexer dedicated to index product models.

in this PR both subscribers point to the same class in which we inject the wanted
indexers.

Regarding the indexing of product model childs, an idea might be to take
advantage of the event system and somehow recursively throw one event
per product model child, so that those are taken care of until the tree
has been fully processed.